FIBER_TASKING_LIB: Suppress loss of data error
If we have more than UINT_MAX threads, there are bigger problems

+++ source/fiber_tasking_lib/task_scheduler.cpp
@@ -244,7 +244,7 @@ void TaskScheduler::Quit() {
 		}
 	}
 
-	FTLJoinThreads(workerThreads.size(), &workerThreads[0]);
+	FTLJoinThreads((uint)workerThreads.size(), &workerThreads[0]);
 
 	for (auto &workerThread : workerThreads) {
 		CloseHandle(workerThread);
